+++ examples/template-hydrogen-default/src/components/Cart.client.jsx
@@ -18,7 +18,7 @@ import {BUTTON_PRIMARY_CLASSES} from './Button.client';
  */
 export default function Cart() {
   const {isCartOpen, closeCart} = useCartUI();
-  const {cartLinesTotalQuantity} = useCart();
+  const {totalQuantity} = useCart();
 
   return (
     <div>
@@ -33,13 +33,11 @@ export default function Cart() {
         <Dialog.Overlay className="fixed z-20 inset-0 bg-gray-50 opacity-75" />
         <div
           className={`absolute flex flex-col md:block z-20 top-0 left-0 right-0 bottom-0 md:top-7 h-full md:left-auto md:right-7 md:bottom-auto md:h-auto md:max-h-[calc(100vh-56px)] bg-gray-50 w-full md:w-[470px] rounded-b-lg shadow-2xl ${
-            cartLinesTotalQuantity === 0
-              ? 'overflow-hidden'
-              : 'overflow-y-scroll'
+            totalQuantity === 0 ? 'overflow-hidden' : 'overflow-y-scroll'
           }`}
         >
           <CartHeader />
-          {cartLinesTotalQuantity === 0 ? (
+          {totalQuantity === 0 ? (
             <CartEmpty />
           ) : (
             <>